requirements.txt
sys
PyQt5.uic
PyQt5.QtSql.QSqlDatabase, PyQt5.QtSql.QSqlTableModel
PyQt5.QtWidgets.QWidget, PyQt5.QtWidgets.QApplication